More robust integration tests, move them to tests folder
